Add more Javadoc
